# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: coach
    description: Coach Meta chat
  - name: analyze
    description: Gameplay analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /coach/chat:
    post:
      operationId: coachChat
      tags: [coach]
      summary: Coach Meta chat
      description: Game-grounded coaching reply for a selected title
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachChatRequest"
      responses:
        "200":
          description: Coach reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachChatResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyze:
    post:
      operationId: analyzeGameplay
      tags: [analyze]
      summary: Analyze gameplay session
      description: Returns a meta-grounded analysis for uploaded gameplay metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CoachChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    CoachChatRequest:
      type: object
      properties:
        gameId:
          type: string
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/CoachChatMessage"
      required:
        - gameId
        - message
    CoachChatResponse:
      type: object
      properties:
        reply:
          type: string
        conceptsUsed:
          type: array
          items:
            type: string
        source:
          type: string
          enum: [llm, knowledge, canned]
      required:
        - reply
        - conceptsUsed
        - source
    AnalyzeRequest:
      type: object
      properties:
        gameId:
          type: string
        fileName:
          type: string
        durationSeconds:
          type: number
        observations:
          type: array
          items:
            type: string
      required:
        - gameId
    AnalyzeResponse:
      type: object
      properties:
        matchRead:
          type: string
        whatYouWereAbusing:
          type: array
          items:
            type: string
        whatTheyWereAbusing:
          type: array
          items:
            type: string
        biggestTell:
          type: string
        staminaEconomy:
          type: string
        scoringBattle:
          type: string
        missedPunishes:
          type: array
          items:
            type: string
        metaAdjustment:
          type: array
          items:
            type: string
        clipEvidence:
          type: array
          items:
            type: string
        summary:
          type: string
        conceptsUsed:
          type: array
          items:
            type: string
        analysisLimits:
          type: string
        confidenceThreshold:
          type: number
        viewerSide:
          type: string
          enum: [left, right, unknown]
        evidenceMode:
          type: string
          enum: [frames, thumbnail, none]
        eventLog:
          type: array
          items:
            type: object
            properties:
              timestampSeconds:
                type: number
              actor:
                type: string
              action:
                type: string
              result:
                type: string
              confidence:
                type: number
              visibleEvidence:
                type: string
        strengths:
          type: array
          items:
            type: string
          description: Deprecated — kept for older clients; prefer tape sections
        weaknesses:
          type: array
          items:
            type: string
          description: Deprecated — kept for older clients; prefer tape sections
      required:
        - matchRead
        - whatYouWereAbusing
        - whatTheyWereAbusing
        - biggestTell
        - staminaEconomy
        - scoringBattle
        - missedPunishes
        - metaAdjustment
        - clipEvidence
        - summary
        - conceptsUsed
        - strengths
        - weaknesses
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error