# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Game-grounded coaching reply for a selected title
 * @summary Coach Meta chat
 */
export const CoachChatBody = zod.object({
  "gameId": zod.string(),
  "message": zod.string(),
  "history": zod.array(zod.object({
  "role": zod.enum(['user', 'assistant']),
  "content": zod.string()
})).optional()
})

export const CoachChatResponse = zod.object({
  "reply": zod.string(),
  "conceptsUsed": zod.array(zod.string()),
  "source": zod.enum(['llm', 'knowledge', 'canned'])
})


/**
 * Returns a meta-grounded analysis for uploaded gameplay metadata
 * @summary Analyze gameplay session
 */
export const AnalyzeGameplayBody = zod.object({
  "gameId": zod.string(),
  "fileName": zod.string().optional(),
  "durationSeconds": zod.number().optional(),
  "observations": zod.array(zod.string()).optional()
})

export const     AnalyzeGameplayResponse = zod.object({
  "matchRead": zod.string(),
  "whatYouWereAbusing": zod.array(zod.string()),
  "whatTheyWereAbusing": zod.array(zod.string()),
  "biggestTell": zod.string(),
  "staminaEconomy": zod.string(),
  "scoringBattle": zod.string(),
  "missedPunishes": zod.array(zod.string()),
  "metaAdjustment": zod.array(zod.string()),
  "clipEvidence": zod.array(zod.string()),
  "summary": zod.string(),
  "conceptsUsed": zod.array(zod.string()),
  "strengths": zod.array(zod.string()),
  "weaknesses": zod.array(zod.string()),
  "analysisLimits": zod.string().optional(),
  "confidenceThreshold": zod.number().optional(),
  "viewerSide": zod.enum(['left', 'right', 'unknown']).optional(),
  "evidenceMode": zod.enum(['frames', 'thumbnail', 'none']).optional()
})